fix(ci): mypy type fixes, black formatting, ruff clean, 83 tests passing

+++ src/services/anthropic.py
@@ -178,7 +178,7 @@ def _calculate_delay_ms(attempt: int, base_delay_ms: int) -> int:
     Returns:
         Delay in milliseconds.
     """
-    return base_delay_ms * (2 ** (attempt - 1))
+    return int(base_delay_ms * (2 ** (attempt - 1)))
 
 
 class _RetryableApiError(Exception):
